Fix indexing error in TwoPhaseDefUseMap
There was an off-by-one in the indexing for TwoPhaseDefUseMap, causing
def-use chains to not be properly visited. We don't use this functionality
much in base, because it's only active for irinterp on functions with
loops that were shown to terminate, which the compiler currently does
not generally have the power to do, but I saw it in some downstream
experiments.