Align meaning for effects and IR flags
This fixes a longstanding todo where the IR_FLAG_EFFECT_FREE flag
actually required both :effect_free and :nothrow. After this PR,
it is equivalent to :effect_free only. The mismatch in meaning here
caused #50311. `Symbol(::String)` is :effect_free, but not :nothrow.
As a result, setting IR_FLAG_EFFECT_FREE on it was not legal. Later,
irinterp did discover that it was nothrow and set IR_FLAG_NOTHROW,
but did not have sufficient information to know that it was also
:effect_free, so it could not set that flag. With this PR,
IR_FLAG_EFFECT_FREE is set early in inference, so once irinterp
discovers IR_FLAG_NOTHROW, the call becomes DCE-eligible as
desired. Fixes #50311.