Flip trust_inference option
I think the time has come to flip this. This was added when the type
system was much less reliable at producing intersections. It is true
that we still have the occasional type sytem bug, but we're already
trusting inference in a bunch of other places. At the same time,
the cost of this has grown in terms of bloated IR needing to
be visited in places like irinterp, so let's flip the bit and
we'll deal with type system bugs the way we usually due.