Remove union penalties for inlining cost (#50429)
I added this code back in #27057, when I first made Union-full
signatures inlineable. The justification was to try to encourage
the union splitting to happen on the outside. However (and I believe
this changed since this code was introduced), these days inference
is in complete control of union splitting and we do not take
inlineability or non-inlineability of the non-unionsplit function
into account when deciding how to inline. As a result, the only
effect of the union split penalties was to prevent inlining of
functions that are not union-split eligible (e.g.
`+(::Vararg{Union{Int, Missing}, 3})`), but are nevertheless cheap
by our inlining metric. There is really no reason not to try to
inline such functions, so delete this logic.