ssair: Correctly handle stmt insertion at end of basic block
In the presence of `attach_after` insertions, we have to be careful to
extend the basic block to include everything up to the last insertion.
We were accounting for "new" nodes (before the compaction point), but
not "pending" nodes (after the compaction point).

Fixes https://github.com/JuliaLang/julia/issues/50379.