Add `--compiled-modules=existing` command line option (#50586)
Occasionally when developing new compiler features, it is not possible
for me to precompile the package I'm developing (because it relies on
the new compiler features, which are only available if Revise'd in).
`--compiled-modules=no` is an option, but for packages with deep
dependency stacks, this is not practical (in addition to slowing down
use of development utilities like Revise, Cthulhu or Plots). Originally
I tried to add a mode to `--compiled-modules` that would avoid using
compiled modules for anything that's dev'ed, but that's a pretty
complicated thing to figure out in the loading code, because you could
have a non-`dev`'ed package that depends on a `dev`'ed package and you'd
want to avoid loading that as well, but we don't technically have
explicit dependency links at the loading level. This sidesteps all that
and just adds a simpler option: `existing`. This option simply uses any
pre-existing cache files if they exist, but refuses to create new ones.
This does effectively the same thing, because the only packages with
stale cache files are usually the ones that I've edited. However, the
semantics are much simpler for loading to implement.