Fix log1p(z) for extremely small |z|. (#50769)
If |z| is extremely small (e.g. `1e-200+5e-175im`), the expression
`log(u)*z/(u-1)` (where `u` is `1 + z`) evaluates to 0, because the
product `log(u)*z` is approximately `z.im*z` when |z| is very small, and
that product underflows to 0. By adding parentheses so that the
expression is `log(u)*(z/(u-1))`, `z/(u - 1)` is evaluated first and the
underflow is avoided.