Add fallback if we have make a weird GC decision. (#50682)
If something odd happens during GC (the PC goes to sleep) or a very big
transient the heuristics might make a bad decision. What this PR
implements is if we try to make our target more than double the one we
had before we fallback to a more conservative method. This fixes the new
issue @vtjnash found in https://github.com/JuliaLang/julia/issues/40644
for me.