Update dependency builds (#50826)
This bumps the build numbers for stdlib and binary dependency JLLs,
updates libssh2 to 1.11.0, libgit2 to 1.6.4, and objconv to 2.53.

Julia's FreeBSD CI has been running on FreeBSD 13.2 for a while, but
until more recently, Yggdrasil was still building FreeBSD binaries using
the 12.2 sysroot. The sysroot was updated to 13.2 and I went through and
rebuilt the dependencies that Julia uses. The updated build numbers
correspond to these rebuilt but otherwise unchanged binaries.

The actual version updates are because libssh2 in Yggdrasil was at
1.11.0 so I left it there (its [release
notes](https://github.com/libssh2/libssh2/releases/tag/libssh2-1.11.0)
suggest it's a safe update), libgit2 had a newer patch version available
and needed to be fixed anyway since the Windows build was broken, and
objconv needed its Yggdrasil build recipe fixed but Elliot's GitHub
mirror of objconv was at 2.53 so I updated to use that.