compiler: Rename :inst -> :stmt (#50868)
As discussed in #49057, this renames the :inst field of `Instruction` to
`:stmt` (`:inst` is still allowed for the moment to ease transition for
out-of-tree users). I'm hoping to get through some more compiler
datastructure cleanup in the 1.11 timeframe, but this is probably one of
the larger ones in terms of number of lines affected.