Remove `export_small_typeof`
I think this might have been part of a scheme to avoid an extra
indirection when this symbol is exposed to the linker. We have a similar
technique used for JIT'd modules, but I'm not sure how it's supposed
to work here between libjulia and libjulia-internal.

On Windows, this is broken for embedding currently, and on Linux this
function was just copying from the `small_typeof` in libjulia to a seem-
ingly unused copy of the symbol in libjulia-internal.

For now, this removes the special machinery and replaces it with a single
copy of the data. If we'd like to optimize this in the future, we might
want to use the existing "ijl_" vs. "jl_" pattern