revert names to the most basic implementation (simply swap export=>public) and publicize `isexported` and `ispublic`
How to get the old behavior? `filter(isexported, names(args...; kw...)`

The set of possible keywods and interacitons is huge and it is better to give folks composable functions than a giant, tangled keyword system that is bound to be eternaly cludgey to maintain backwards compatability.

Documents but does not publicize `isdeprecated` and `isbindingresolved`.