timing: Check `jl_current_task` is ready before emitting timing events
We sometimes spawn imaging-dedicated threads for running LLVM that do
not have `jl_current_task` setup. Since we rely on the thread-local
storage stack for our built-in profiling, we will segfault if we try
to emit any events on these threads.

A better long-term fix is probably to setup `jl_current_task` et al
on these threads so that we can track them as usual, but this avoids
crashes for now.