slot2ssa: Insert PiNodes based on `bb_vartables`
This is the last usage of TypedSlot (we should be clear now to remove
that step from the end of type inference now).

Compared to the prior scheme, this now inserts PiNodes at every PhiNode
corresponding to a given slot, rather than at each load. In principle,
that should mean that the types we compute across PhiNodes are more
precise than before.

I believe that means that all slot uses should be at least as precise
as before, but I haven't completely convinced myself yet.